Implement passing Tunnistamo API tokens downstream
This gateway has the following flow for authentication:
- when a user logs in with Tunnistamo, they get an id_token and
  and access token
- with the access token, they can ask Tunnistamo for API tokens
  for the API scopes, the frontend client is allowed to access
- frontend needs to pass the entire json with those API tokens
  to the gateway in the "Api-Tokens" header
- the gateway server parses that header and uses appropriate
  token, when calling each downstream API

This way, we can keep the current Tunnistamo auth flow intact,
where one API token can only be used for one API. We also
currently do not want to implement user authentication directly
at the gateway level, as the apollo-federation docs suggest,
since 1) we don't have readymade solutions for this in TS/JS,
2) we haven't fully investigated the security implications of
this approach (what a breach at the gateway would entail).
Maybe in the future, if these two issues get required attention,
gateway will handle the auth, but not now.